+++ oldtests/bootstrap.php
@@ -1,4 +1,0 @@
-<?php
-
-// use Composer's autoloader
-require (__DIR__ . "/../vendor/autoload.php");