+++ tests/HashedVersions/Parsers/ParseHashedVersionTest.php
@@ -34,26 +34,26 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Versions/VersionNumbers
+ * @package   Versions/HashedVersions/Parsers
  * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
  * @copyright 2015-present Ganbaro Digital Ltd www.ganbarodigital.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://code.ganbarodigital.com/php-versions
  */
 
-namespace GanbaroDigital\Versions\VersionNumbers\Maps;
+namespace GanbaroDigital\Versions\HashedVersions\Parsers;
 
-use GanbaroDigital\Versions\HashedVersions\Operators\CompareHashedVersions;
-use GanbaroDigital\Versions\HashedVersions\Values\HashedVersion;
-use GanbaroDigital\Versions\SemanticVersions\Operators\CompareSemanticVersions;
-use GanbaroDigital\Versions\SemanticVersions\Parsers\ParseSemanticVersion;
-use GanbaroDigital\Versions\VersionNumbers\Values\VersionNumber;
+require_once(__DIR__ . '/../../Datasets/HashedVersionDatasets.php');
+
 use PHPUnit_Framework_TestCase;
+use GanbaroDigital\Versions\Datasets\HashedVersionDatasets;
+use GanbaroDigital\Versions\VersionNumbers\Parsers\VersionParser;
+use stdClass;
 
 /**
- * @coversDefaultClass GanbaroDigital\Versions\VersionNumbers\Maps\GetVersionComparitor
+ * @coversDefaultClass GanbaroDigital\Versions\HashedVersions\Parsers\ParseHashedVersion
  */
-class GetVersionComparitorTest extends PHPUnit_Framework_TestCase
+class ParseHashedVersionTest extends PHPUnit_Framework_TestCase
 {
     /**
      * @coversNothing
@@ -66,41 +66,61 @@ class GetVersionComparitorTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $obj = new GetVersionComparitor;
+        $obj = new ParseHashedVersion;
 
         // ----------------------------------------------------------------
         // test the results
 
-        $this->assertTrue($obj instanceof GetVersionComparitor);
+        $this->assertTrue($obj instanceof ParseHashedVersion);
+    }
+
+    /**
+     * @coversNothing
+     */
+    public function testIsInstanceOfVersionParser()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $obj = new ParseHashedVersion;
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue($obj instanceof VersionParser);
     }
 
     /**
      * @covers ::__invoke
-     * @dataProvider provideVersionsToTest
+     * @dataProvider provideHashedVersions
      */
-    public function testCanUseAsObject($version, $expectedResult)
+    public function testCanUseAsObject($version)
     {
         // ----------------------------------------------------------------
         // setup your test
 
-        $obj = new GetVersionComparitor;
+        $parser = new ParseHashedVersion;
 
         // ----------------------------------------------------------------
         // perform the change
 
-        $actualResult = $obj($version);
+        $actualResult = $parser($version);
 
         // ----------------------------------------------------------------
         // test the results
 
-        $this->assertTrue($actualResult instanceof $expectedResult);
+        $this->assertEquals($version, (string)$actualResult);
     }
 
     /**
-     * @covers ::matching
-     * @dataProvider provideVersionsToTest
+     * @covers ::from
+     * @covers ::getRegex
+     * @dataProvider provideHashedVersions
      */
-    public function testCanCallStatically($version, $expectedResult)
+    public function testCanCallStatically($version)
     {
         // ----------------------------------------------------------------
         // setup your test
@@ -108,117 +128,75 @@ class GetVersionComparitorTest extends PHPUnit_Framework_TestCase
         // ----------------------------------------------------------------
         // perform the change
 
-        $actualResult = GetVersionComparitor::matching($version);
+        $actualResult = ParseHashedVersion::from($version);
 
         // ----------------------------------------------------------------
         // test the results
 
-        $this->assertTrue($actualResult instanceof $expectedResult);
+        $this->assertEquals($version, (string)$actualResult);
     }
 
     /**
-     * @covers ::matching
-     * @expectedException GanbaroDigital\Versions\Exceptions\E4xx_UnsupportedType
+     * @covers ::from
+     * @dataProvider provideNonStrings
+     * @expectedException GanbaroDigital\Versions\Exceptions\E4xx_NotAVersionString
      */
-    public function testThrowsExceptionIfNoComparitorExists()
+    public function testRejectsNonStrings($data)
     {
         // ----------------------------------------------------------------
         // setup your test
 
-        $version = new FakeVersionNumber;
+        // ----------------------------------------------------------------
+        // perform the change
+
+        ParseHashedVersion::from($data);
+    }
+
+    /**
+     * @covers ::from
+     * @dataProvider provideBadVersionStrings
+     * @expectedException GanbaroDigital\Versions\Exceptions\E4xx_BadVersionString
+     */
+    public function testRejectsBadVersionStrings($data)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
 
         // ----------------------------------------------------------------
         // perform the change
 
-        GetVersionComparitor::matching($version);
+        ParseHashedVersion::from($data);
     }
 
 
-    public function provideVersionsToTest()
+    public function provideHashedVersions()
+    {
+        $retval = [];
+        foreach (HashedVersionDatasets::getVersionNumberDataset() as $data) {
+            $retval[] = [ $data[0] ];
+        }
+
+        return $retval;
+    }
+
+    public function provideNonStrings()
     {
         return [
-            [ ParseSemanticVersion::from("1.0"), CompareSemanticVersions::class ],
-            [ new HashedVersion("abcdef"), CompareHashedVersions::class ],
+            [ null ],
+            [ true ],
+            [ false ],
+            [ [] ],
+            [ function() { return null; } ],
+            [ 3.1415927 ],
+            [ 0 ],
+            [ 1000 ],
+            [ new stdClass ],
+            [ STDIN ],
         ];
     }
-}
 
-class FakeVersionNumber implements VersionNumber
-{
-    public function getVersion()
+    public function provideBadVersionStrings()
     {
-        return null;
-    }
-
-    public function getMajor()
-    {
-        return null;
-    }
-
-    public function getMinor()
-    {
-        return null;
-    }
-
-    public function hasPatchLevel()
-    {
-        return false;
-    }
-
-    public function getPatchLevel()
-    {
-        return null;
-    }
-
-    public function hasPreRelease()
-    {
-        return false;
-    }
-
-    public function getPreRelease()
-    {
-        return null;
-    }
-
-    public function hasBuildNumber()
-    {
-        return false;
-    }
-
-    public function getBuildNumber()
-    {
-        return null;
-    }
-
-    // ==================================================================
-    //
-    // Calculated properties go here
-    //
-    // ------------------------------------------------------------------
-
-    public function getApproximateUpperBoundary()
-    {
-        return null;
-    }
-
-    public function getCompatibleUpperBoundary()
-    {
-        return null;
-    }
-
-    // ==================================================================
-    //
-    // Type convertors go here
-    //
-    // ------------------------------------------------------------------
-
-    public function __toString()
-    {
-        return '';
-    }
-
-    public function toArray()
-    {
-        return [];
+        return HashedVersionDatasets::getBadVersionNumberDataset();
     }
 }